ref(ui): moved load-dayjs-locale file
